backend/requirements.txt
gradio>=4.0.0,<5.0.0
Pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<2.0.0
psd-tools>=2.0.0,<3.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0,<5.0.0
simple-lama-inpainting>=0.1.0
